Move genOverlayICPS() to a better place, reducing imports in upgrade.go.
